tresor_backend/requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
pillow==10.3.0
python-dateutil==2.9.0.post0
sqlparse==0.5.0
pytz==2024.1
psycopg2-binary==2.9.9 